requirements.txt
pydantic==2.10.6
python-dotenv
langchain
langchain_chroma
langchain_community
pypdf
langchain-text-splitters
langchain-openai
chromadb
sentence-transformers
langchain_huggingface
faiss-cpu
streamlit
watchdog
langchain_groq
fastapi
uvicorn
langserve[server]
sse_starlette
langgraph
langgraph-checkpoint-sqlite
boto3
langchain-aws>=0.1.0
langchain-ollama>=0.1.0
aiosqlite
rank_bm25
jq